# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductStockStatus = typeof ProductStockStatus[keyof typeof ProductStockStatus];


export const ProductStockStatus = {
  safe: 'safe',
  low: 'low',
  out: 'out',
} as const;

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  sku?: string | null;
  category: string;
  purchasePrice: number;
  salePrice: number;
  stock: number;
  lowStockThreshold: number;
  unit: string;
  stockStatus: ProductStockStatus;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  sku?: string | null;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  purchasePrice: number;
  /** @minimum 0 */
  salePrice: number;
  /** @minimum 0 */
  stock: number;
  /** @minimum 0 */
  lowStockThreshold: number;
  /** @minLength 1 */
  unit: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  sku?: string | null;
  /** @minLength 1 */
  category?: string;
  /** @minimum 0 */
  purchasePrice?: number;
  /** @minimum 0 */
  salePrice?: number;
  /** @minimum 0 */
  stock?: number;
  /** @minimum 0 */
  lowStockThreshold?: number;
  /** @minLength 1 */
  unit?: string;
}

export interface Sale {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  total: number;
  createdAt: string;
}

export interface SaleInput {
  /** @minimum 1 */
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface RevenuePoint {
  label: string;
  value: number;
}

export interface TopProduct {
  name: string;
  quantity: number;
  revenue: number;
}

export interface DashboardSummary {
  revenueToday: number;
  revenueMonth: number;
  salesToday: number;
  productCount: number;
  lowStockCount: number;
  weeklyRevenue: RevenuePoint[];
  topProducts: TopProduct[];
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  sale: 'sale',
  product: 'product',
  stock: 'stock',
} as const;

export interface Activity {
  id: number;
  type: ActivityType;
  title: string;
  description: string;
  createdAt: string;
}

export type RestockRecommendationUrgency = typeof RestockRecommendationUrgency[keyof typeof RestockRecommendationUrgency];


export const RestockRecommendationUrgency = {
  urgent: 'urgent',
  soon: 'soon',
  planned: 'planned',
} as const;

export interface RestockRecommendation {
  productId: number;
  productName: string;
  currentStock: number;
  suggestedQuantity: number;
  avgDailySales: number;
  urgency: RestockRecommendationUrgency;
  reason: string;
}

export type ListProductsParams = {
search?: string;
stockStatus?: ListProductsStockStatus;
};

export type ListProductsStockStatus = typeof ListProductsStockStatus[keyof typeof ListProductsStockStatus];


export const ListProductsStockStatus = {
  all: 'all',
  low: 'low',
  out: 'out',
} as const;

export type ListSalesParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};