# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RevenuePoint } from './revenuePoint';
import type { TopProduct } from './topProduct';

export interface DashboardSummary {
  revenueToday: number;
  revenueMonth: number;
  salesToday: number;
  productCount: number;
  lowStockCount: number;
  weeklyRevenue: RevenuePoint[];
  topProducts: TopProduct[];
}