# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStockStatus } from './productStockStatus';

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  sku?: string | null;
  category: string;
  purchasePrice: number;
  salePrice: number;
  stock: number;
  lowStockThreshold: number;
  unit: string;
  stockStatus: ProductStockStatus;
  createdAt: Date;
  updatedAt: Date;
}