# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  sku?: string | null;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  purchasePrice: number;
  /** @minimum 0 */
  salePrice: number;
  /** @minimum 0 */
  stock: number;
  /** @minimum 0 */
  lowStockThreshold: number;
  /** @minLength 1 */
  unit: string;
}