# lib/api-zod/src/generated/types/restockRecommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RestockRecommendationUrgency } from './restockRecommendationUrgency';

export interface RestockRecommendation {
  productId: number;
  productName: string;
  currentStock: number;
  suggestedQuantity: number;
  avgDailySales: number;
  urgency: RestockRecommendationUrgency;
  reason: string;
}