# lib/api-zod/src/generated/types/saleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaleInput {
  /** @minimum 1 */
  productId: number;
  /** @minimum 1 */
  quantity: number;
}